// src/openapi-request/models/CreateBranchDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateBranchDto = {
    _oid?: string;
    branchName: string;
    branchCode: string;
    branchAddress: string;
    branchPhoneNumber: string;
    branchEmail: string;
    branchWebsite: string;
    branchDescription: string;
    branchType: string;
    branchStatus: string;
    managerId?: number;
};